+++ tests/grd/test_commands.py
@@ -1,4 +1,5 @@
 import os
+from pathlib import Path
 from tempfile import TemporaryDirectory
 from typing import Callable, List
 
@@ -9,7 +10,6 @@ from pystac.utils import is_absolute_href
 from stactools.testing.cli_test import CliTestCase
 
 from stactools.sentinel1.commands import create_sentinel1_command
-from stactools.sentinel1.grd import stac
 from stactools.sentinel1.grd.constants import SENTINEL_POLARIZATIONS
 from tests import test_data
 
@@ -19,13 +19,23 @@ class CreateItemTest(CliTestCase):
         return [create_sentinel1_command]
 
     def test_create_collection(self) -> None:
-        collection = stac.create_collection()
-        assert isinstance(collection, pystac.Collection)
+        with TemporaryDirectory() as tmp_dir:
+            tmp_path = Path(tmp_dir)
+            destination = tmp_path / "sentinel1-grd.json"
+            cmd = ["sentinel1", "grd", "create-collection", f"{tmp_path}"]
+            result = self.run_command(cmd)
 
-    def test_validate_collection(self) -> None:
-        collection = stac.create_collection()
-        collection.normalize_hrefs("./")
-        collection.validate()
+            assert result.exit_code == 0, "\n{}".format(result.output)
+            paths = [p for p in tmp_path.iterdir() if p.suffix == ".json"]
+
+            assert len(paths) == 1
+            collection = pystac.Collection.from_file(str(destination))
+
+            assert collection.id == "sentinel1-grd"
+            # collection.set_self_href(
+            #    str(destination)
+            # )  # Must set the self reference to pass validation
+            collection.validate()
 
     def test_create_item(self) -> None:
         item_id = "S1A_IW_GRDH_1SDV_20210809T173953_20210809T174018_039156_049F13_6FF8"